requirements.txt
numpy~=1.25.1
opencv-python~=4.8.0.74
ultralytics~=8.0.135
shapely~=2.0.1
py-cpuinfo
python-dotenv~=1.0.0
